requirements.txt
Flask
flask-mysqldb
